+++ sources/common/time/time.h
@@ -6,11 +6,11 @@ class Time {
 public:
     Time() = default;
 
-    static Time FromSeconds(float seconds);
+    static Time FromSeconds(double seconds);
     static Time FromMilliseconds(int64_t ms);
     static Time FromMicroseconds(int64_t us);
 
-    float AsSeconds() const;
+    double AsSeconds() const;
     int64_t AsMilliseconds() const;
     int64_t AsMicroseconds() const;
     bool IsZero() const;
@@ -38,17 +38,17 @@ Time& operator+=(Time& left, Time right);
 Time operator-(Time left, Time right);
 Time& operator-=(Time& left, Time right);
 
-Time operator*(Time left, float right);
-Time operator*(float left, Time right);
-Time& operator*=(Time& left, float right);
+Time operator*(Time left, double right);
+Time operator*(double left, Time right);
+Time& operator*=(Time& left, double right);
 
 Time operator*(Time left, uint64_t right) = delete;
 Time operator*(uint64_t left, Time right) = delete;
 Time& operator*=(Time& left, uint64_t right) = delete;
 
-Time operator/(Time left, float right);
-Time& operator/=(Time& left, float right);
-float operator/(Time left, Time right);
+Time operator/(Time left, double right);
+Time& operator/=(Time& left, double right);
+double operator/(Time left, Time right);
 
 Time operator/(Time left, uint64_t right) = delete;
 Time& operator/=(Time& left, uint64_t right) = delete;
